cge-system/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
passlib[bcrypt]
python-jose[cryptography]
rapidfuzz>=3.0.0
pytesseract
Pillow
python-multipart
numpy
paddlepaddle==2.6.1
paddleocr==2.7.3
opencv-python-headless
google-cloud-vision>=3.5.0
twilio>=8.0.0
face_recognition
dlib
